adding toggle buttons in write code
Can't find an actual toggle button in PHP so I used a regular button
instead. It was the closest thing and the toggle buttons in access have
a caption on it. Actual closest thing in html is radio button but it
cant have a caption on it unless you make a new label to the left or
right...