Introducing Python and Converting Rowsources
You guy may hate me for this, but I'm including Python scripts in the
program. It's pretty simple to use. Download http://ironpython.net/ and
you may need python 2.7.8 but you might not. After that, it should work.

IronPython allows us to transfer information through variables and run
scripts in both languages. Look at test.py to see how it works.

I used IronPython in QueryFind.py to go through the .txt files and find
the correct rowsource for each listbox and combobox and after getting
the correct query in the result variable, I grabbed it in PageInfo.cs
and used it to write the PHP we need.